requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
httpx
faiss-cpu
numpy
Pillow
crewai
